requirements.txt
behave==1.2.5
Django==1.8.4
django-model-utils==2.3.1
EasyProcess==0.1.9
jsonfield==1.0.3
parse==1.6.6
parse-type==0.3.4
PyVirtualDisplay==0.1.5
selenium==2.47.1
six==1.9.0
